int64 from int32

+++ MultiTrain/test/utils_test.py
@@ -44,7 +44,7 @@ def test_metrics(custom_metric, expected_exception):
 def test_cat_encoder(sample_dataframe):
     encoded_df = _cat_encoder(sample_dataframe, auto_cat_encode=True)
     assert 'cat_col' in encoded_df.columns
-    assert encoded_df['cat_col'].dtype == 'int32'
+    assert encoded_df['cat_col'].dtype == 'int64'
 
 def test_manual_encoder(sample_dataframe):
     manual_encode = {'label': ['cat_col'], 'onehot': []}
